Fixed flexible content, also moved helper methods for repeater & flexible content to the repository to keep it DRY
